Refactor SearchworksItem::..::RequestStatus to top-level ItemStatus, and only persist non-default statuses
